feat(quadraui): lift GTK editor rasteriser (#276 stage 1D)
Phase C Stage 1D completes the editor primitive lift. The GTK paint
moves into `quadraui::gtk::draw_editor` (verbatim port from
`src/gtk/draw::draw_window`); vimcode's `draw_window` collapses to a
~25-line delegator that routes through `render::to_q_editor` and
`q_theme()`. Net -700+ LOC removed from `src/gtk/draw.rs`.

Changes:

- `quadraui/src/gtk/editor.rs` (new, 776 LOC) — verbatim port of
  `draw_window` body. Selection paints BEFORE text (Cairo painter
  order), distinct from TUI (which paints after text via cell-bg
  modification). Wavy diagnostic underlines, dotted spell underlines,
  alpha-rect cursor, semi-transparent bracket-match. Internal helpers
  `build_pango_attrs` and `draw_visual_selection` are private to this
  module. Scrollbars are NOT painted by this rasteriser (GTK paints
  them outside `draw_window` via `draw_h_scrollbars`; vertical
  scrollbar is a native gtk4 widget). Per-window status line is
  excluded (already lifted, Session 241).
- `src/gtk/draw.rs::draw_window` collapses to delegator. Old body +
  `build_pango_attrs` + `draw_visual_selection` removed.
- `src/gtk/mod.rs` unused-import cleanup (DiagnosticSeverity,
  GitLineStatus, CursorShape, SelectionKind, SelectionRange,
  StyledSpan).

vimcode `--no-default-features` + GTK build + clippy clean.
quadraui `--features tui --features gtk` 290 tests pass. Full
vimcode test suite green (1950 lib + integration). kubeui +
kubeui-gtk consumers build clean.

Stage 1 complete. The editor primitive is now lifted with both TUI
and GTK rasterisers; B.6 (Win-GUI rebuild on quadraui) is unblocked.

Closes #276.
