+++ quadraui/src/gtk/mod.rs
@@ -28,6 +28,7 @@ mod activity_bar;
 pub mod backend;
 mod context_menu;
 mod dialog;
+mod editor;
 pub mod events;
 mod find_replace;
 mod form;
@@ -50,6 +51,7 @@ pub use activity_bar::{draw_activity_bar, ACTIVITY_ROW_PX};
 pub use backend::GtkBackend;
 pub use context_menu::draw_context_menu;
 pub use dialog::draw_dialog;
+pub use editor::draw_editor;
 pub use find_replace::draw_find_replace;
 pub use form::{draw_form, draw_settings_chrome};
 pub use list::draw_list;
