+++ src/gtk/draw.rs
@@ -1081,6 +1081,18 @@ pub(super) fn draw_breadcrumb_bar(
     });
 }
 
+/// Render one editor window (pane) onto `cr`.
+///
+/// Phase C Stage 1D (#276) collapsed the body of this function to a
+/// thin delegator. The actual paint code lives in
+/// `quadraui::gtk::draw_editor`, fed by `render::to_q_editor` (the
+/// boundary adapter that converts the engine-side `RenderedWindow`
+/// IR into the cross-backend `quadraui::Editor` primitive). GTK
+/// scrollbars are painted elsewhere (`draw_h_scrollbars` for
+/// horizontal, native gtk4 widgets for vertical) — preserved by this
+/// delegator since the rasteriser explicitly excludes scrollbar paint
+/// on GTK. The per-window status line is also painted separately (it
+/// was lifted in Session 241).
 #[allow(clippy::too_many_arguments)]
 pub(super) fn draw_window(
     cr: &Context,
@@ -1091,717 +1103,17 @@ pub(super) fn draw_window(
     char_width: f64,
     line_height: f64,
 ) {
-    let rect = &rw.rect;
-
-    // Gutter pixel width
-    let gutter_width = rw.gutter_char_width as f64 * char_width;
-
-    // Apply horizontal scroll offset
-    let h_scroll_offset = rw.scroll_left as f64 * char_width;
-    let text_x_offset = rect.x + gutter_width - h_scroll_offset;
-
-    // Window background
-    let bg = if rw.show_active_bg {
-        theme.active_background
-    } else {
-        theme.background
-    };
-    let (br, bg_g, bb) = bg.to_cairo();
-    cr.set_source_rgb(br, bg_g, bb);
-    cr.rectangle(rect.x, rect.y, rect.width, rect.height);
-    cr.fill().ok();
-
-    // Cursorline / Diff / DAP stopped-line background (drawn before selection so selection is on top)
-    for (view_idx, rl) in rw.lines.iter().enumerate() {
-        let y = rect.y + view_idx as f64 * line_height;
-        let bg_color = if rl.is_dap_current {
-            Some(theme.dap_stopped_bg)
-        } else if let Some(diff_status) = rl.diff_status {
-            use crate::core::engine::DiffLine;
-            match diff_status {
-                DiffLine::Added => Some(theme.diff_added_bg),
-                DiffLine::Removed => Some(theme.diff_removed_bg),
-                DiffLine::Padding => Some(theme.diff_padding_bg),
-                DiffLine::Same => None,
-            }
-        } else if rl.is_current_line && rw.is_active && rw.cursorline {
-            Some(theme.cursorline_bg)
-        } else {
-            None
-        };
-        if let Some(color) = bg_color {
-            let (dr, dg, db) = color.to_cairo();
-            cr.set_source_rgb(dr, dg, db);
-            cr.rectangle(rect.x, y, rect.width, line_height);
-            cr.fill().ok();
-        }
-    }
-
-    // Visual selection highlight (drawn before text so text renders on top)
-    if let Some(sel) = &rw.selection {
-        draw_visual_selection(
-            cr,
-            layout,
-            sel,
-            &rw.lines,
-            rect,
-            line_height,
-            rw.scroll_top,
-            text_x_offset,
-            theme.selection,
-            theme.selection_alpha,
-        );
-    }
-
-    // Extra selections (Ctrl+D multi-cursor word highlights)
-    for esel in &rw.extra_selections {
-        draw_visual_selection(
-            cr,
-            layout,
-            esel,
-            &rw.lines,
-            rect,
-            line_height,
-            rw.scroll_top,
-            text_x_offset,
-            theme.selection,
-            theme.selection_alpha,
-        );
-    }
-
-    // Yank highlight (brief flash after yank)
-    if let Some(yh) = &rw.yank_highlight {
-        draw_visual_selection(
-            cr,
-            layout,
-            yh,
-            &rw.lines,
-            rect,
-            line_height,
-            rw.scroll_top,
-            text_x_offset,
-            theme.yank_highlight_bg,
-            theme.yank_highlight_alpha,
-        );
-    }
-
-    // Render gutter (bp marker + git marker + fold indicators + optional line numbers)
-    if rw.gutter_char_width > 0 {
-        for (view_idx, rl) in rw.lines.iter().enumerate() {
-            let y = rect.y + view_idx as f64 * line_height;
-
-            // Track how many left-aligned marker chars have been rendered.
-            let mut char_offset = 0usize;
-
-            // Breakpoint column — leftmost when any breakpoints/session active.
-            if rw.has_breakpoints {
-                let bp_ch: String = rl.gutter_text.chars().take(1).collect();
-                let bp_color = if rl.is_dap_current || rl.is_breakpoint {
-                    theme.diagnostic_error
-                } else {
-                    theme.line_number_fg
-                };
-                layout.set_text(&bp_ch);
-                layout.set_attributes(None);
-                let (br, bg_c, bb) = bp_color.to_cairo();
-                cr.set_source_rgb(br, bg_c, bb);
-                cr.move_to(rect.x + 3.0, y);
-                pangocairo::show_layout(cr, layout);
-                char_offset += 1;
-            }
-
-            // Git marker column.
-            if rw.has_git_diff {
-                let git_ch: String = rl.gutter_text.chars().skip(char_offset).take(1).collect();
-                let git_color = match rl.git_diff {
-                    Some(GitLineStatus::Added) => theme.git_added,
-                    Some(GitLineStatus::Modified) => theme.git_modified,
-                    Some(GitLineStatus::Deleted) => theme.git_deleted,
-                    None => theme.line_number_fg,
-                };
-                layout.set_text(&git_ch);
-                layout.set_attributes(None);
-                let (gr, gg, gb) = git_color.to_cairo();
-                cr.set_source_rgb(gr, gg, gb);
-                cr.move_to(rect.x + char_offset as f64 * char_width + 3.0, y);
-                pangocairo::show_layout(cr, layout);
-                char_offset += 1;
-
-                // Fold+numbers portion right-aligned.
-                let rest: String = rl.gutter_text.chars().skip(char_offset).collect();
-                layout.set_text(&rest);
-                layout.set_attributes(None);
-            } else if char_offset > 0 {
-                // bp column only — rest is fold+numbers.
-                let rest: String = rl.gutter_text.chars().skip(char_offset).collect();
-                layout.set_text(&rest);
-                layout.set_attributes(None);
-            } else {
-                // No marker columns.
-                layout.set_text(&rl.gutter_text);
-                layout.set_attributes(None);
-            }
-
-            let (num_width, _) = layout.pixel_size();
-            let num_x = rect.x + gutter_width - num_width as f64 - char_width + 3.0;
-
-            let num_color = if rw.is_active && rl.is_current_line {
-                theme.line_number_active_fg
-            } else {
-                theme.line_number_fg
-            };
-            let (nr, ng, nb) = num_color.to_cairo();
-            cr.set_source_rgb(nr, ng, nb);
-            cr.move_to(num_x, y);
-            pangocairo::show_layout(cr, layout);
-
-            // Diagnostic gutter icon (colored dot at leftmost gutter position)
-            if let Some(severity) = rw.diagnostic_gutter.get(&rl.line_idx) {
-                let diag_color = match severity {
-                    DiagnosticSeverity::Error => theme.diagnostic_error,
-                    DiagnosticSeverity::Warning => theme.diagnostic_warning,
-                    DiagnosticSeverity::Information => theme.diagnostic_info,
-                    DiagnosticSeverity::Hint => theme.diagnostic_hint,
-                };
-                let (dr, dg, db) = diag_color.to_cairo();
-                cr.set_source_rgb(dr, dg, db);
-                let dot_r = line_height * 0.2;
-                let dot_cx = rect.x + 3.0 + dot_r;
-                let dot_cy = y + line_height * 0.5;
-                cr.arc(dot_cx, dot_cy, dot_r, 0.0, 2.0 * std::f64::consts::PI);
-                cr.fill().ok();
-            } else if !rl.is_wrap_continuation && rw.code_action_lines.contains(&rl.line_idx) {
-                // Code action lightbulb gutter icon
-                let (lr, lg, lb) = theme.lightbulb.to_cairo();
-                cr.set_source_rgb(lr, lg, lb);
-                let bulb_layout = layout.clone();
-                bulb_layout.set_text(icons::LIGHTBULB.nerd);
-                cr.move_to(rect.x + 1.0, y);
-                pangocairo::show_layout(cr, &bulb_layout);
-            }
-        }
-    } // end gutter rendering block
-
-    // Clip text area (excluding gutter)
-    cr.save().ok();
-    cr.rectangle(
-        rect.x + gutter_width,
-        rect.y,
-        rect.width - gutter_width,
-        rect.height,
+    let editor = render::to_q_editor(rw);
+    let q_theme = super::quadraui_gtk::q_theme(theme);
+    quadraui::gtk::draw_editor(
+        cr,
+        layout,
+        font_metrics,
+        &editor,
+        &q_theme,
+        char_width,
+        line_height,
     );
-    cr.clip();
-
-    // Render each visible line
-    for (view_idx, rl) in rw.lines.iter().enumerate() {
-        let y = rect.y + view_idx as f64 * line_height;
-
-        layout.set_text(&rl.raw_text);
-
-        let attrs = build_pango_attrs(&rl.spans);
-        layout.set_attributes(Some(&attrs));
-
-        let (fr, fg_g, fb) = theme.foreground.to_cairo();
-        cr.set_source_rgb(fr, fg_g, fb);
-        cr.move_to(text_x_offset, y);
-        pangocairo::show_layout(cr, layout);
-
-        // Ghost continuation lines — full line drawn in ghost colour.
-        if rl.is_ghost_continuation {
-            if let Some(ghost) = &rl.ghost_suffix {
-                let (gr, gg, gb) = theme.ghost_text_fg.to_cairo();
-                cr.set_source_rgb(gr, gg, gb);
-                cr.move_to(text_x_offset, y);
-                layout.set_text(ghost);
-                layout.set_attributes(None);
-                pangocairo::show_layout(cr, layout);
-            }
-        }
-
-        // Inline annotation / virtual text (e.g. git blame)
-        if let Some(ann) = &rl.annotation {
-            let text_pixel_width = layout.pixel_size().0 as f64;
-            let ann_x = text_x_offset + text_pixel_width + char_width * 2.0;
-            let (ar, ag, ab) = theme.annotation_fg.to_cairo();
-            cr.set_source_rgb(ar, ag, ab);
-            cr.move_to(ann_x, y);
-            layout.set_text(ann);
-            layout.set_attributes(None);
-            pangocairo::show_layout(cr, layout);
-        }
-
-        // Indent guides: thin vertical lines at each guide column
-        if !rl.indent_guides.is_empty() {
-            cr.set_line_width(1.0);
-            for &guide_col in &rl.indent_guides {
-                let is_active = rw.active_indent_col == Some(guide_col);
-                let (gr, gg, gb) = if is_active {
-                    theme.indent_guide_active_fg.to_cairo()
-                } else {
-                    theme.indent_guide_fg.to_cairo()
-                };
-                cr.set_source_rgb(gr, gg, gb);
-                let gx = text_x_offset + guide_col as f64 * char_width;
-                cr.move_to(gx, y);
-                cr.line_to(gx, y + line_height);
-                cr.stroke().ok();
-            }
-        }
-
-        // Color columns: tinted background rectangle at each column
-        if !rl.colorcolumns.is_empty() {
-            let (cr2, cg, cb) = theme.colorcolumn_bg.to_cairo();
-            cr.set_source_rgb(cr2, cg, cb);
-            for &cc_col in &rl.colorcolumns {
-                let cx = text_x_offset + cc_col as f64 * char_width;
-                cr.rectangle(cx, y, char_width, line_height);
-                cr.fill().ok();
-            }
-        }
-
-        // Bracket match highlighting
-        for &(bm_view_line, bm_col) in &rw.bracket_match_positions {
-            if bm_view_line == view_idx {
-                let (br, bg_c, bb) = theme.bracket_match_bg.to_cairo();
-                cr.set_source_rgba(br, bg_c, bb, 0.6);
-                let bx = text_x_offset + bm_col as f64 * char_width;
-                cr.rectangle(bx, y, char_width, line_height);
-                cr.fill().ok();
-            }
-        }
-
-        // Restore layout to match rendered text (needed for correct
-        // index_to_pos when font_scale != 1.0, e.g. markdown headings).
-        layout.set_text(&rl.raw_text);
-        let line_attrs = build_pango_attrs(&rl.spans);
-        layout.set_attributes(Some(&line_attrs));
-
-        // Diagnostic underlines (wavy squiggles)
-        for dm in &rl.diagnostics {
-            let diag_color = match dm.severity {
-                DiagnosticSeverity::Error => theme.diagnostic_error,
-                DiagnosticSeverity::Warning => theme.diagnostic_warning,
-                DiagnosticSeverity::Information => theme.diagnostic_info,
-                DiagnosticSeverity::Hint => theme.diagnostic_hint,
-            };
-            let (dr, dg, db) = diag_color.to_cairo();
-            cr.set_source_rgb(dr, dg, db);
-            cr.set_line_width(1.0);
-
-            let start_byte = rl
-                .raw_text
-                .char_indices()
-                .nth(dm.start_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-            let end_byte = rl
-                .raw_text
-                .char_indices()
-                .nth(dm.end_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-
-            let start_pos = layout.index_to_pos(start_byte as i32);
-            let end_pos = layout.index_to_pos(end_byte as i32);
-            let x0 = text_x_offset + start_pos.x() as f64 / pango::SCALE as f64;
-            let x1 = text_x_offset + end_pos.x() as f64 / pango::SCALE as f64;
-            let underline_y = y + line_height - 2.0;
-
-            // Draw wavy underline
-            let wave_h = 1.5;
-            let wave_len = 4.0;
-            cr.move_to(x0, underline_y);
-            let mut wx = x0;
-            let mut up = true;
-            while wx < x1 {
-                let next_x = (wx + wave_len).min(x1);
-                let cy = if up {
-                    underline_y - wave_h
-                } else {
-                    underline_y + wave_h
-                };
-                cr.curve_to(
-                    wx + (next_x - wx) * 0.5,
-                    cy,
-                    wx + (next_x - wx) * 0.5,
-                    cy,
-                    next_x,
-                    underline_y,
-                );
-                wx = next_x;
-                up = !up;
-            }
-            cr.stroke().ok();
-        }
-
-        // Spell error underlines (dotted underline in spell_error color)
-        for sm in &rl.spell_errors {
-            let (sr, sg, sb) = theme.spell_error.to_cairo();
-            cr.set_source_rgb(sr, sg, sb);
-            cr.set_line_width(1.0);
-
-            let start_byte = rl
-                .raw_text
-                .char_indices()
-                .nth(sm.start_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-            let end_byte = rl
-                .raw_text
-                .char_indices()
-                .nth(sm.end_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-
-            let start_pos = layout.index_to_pos(start_byte as i32);
-            let end_pos = layout.index_to_pos(end_byte as i32);
-            let x0 = text_x_offset + start_pos.x() as f64 / pango::SCALE as f64;
-            let x1 = text_x_offset + end_pos.x() as f64 / pango::SCALE as f64;
-            let underline_y = y + line_height - 2.0;
-
-            // Draw dotted underline
-            let dot_spacing = 3.0;
-            let mut dx = x0;
-            while dx < x1 {
-                cr.rectangle(dx, underline_y, 1.0, 1.0);
-                dx += dot_spacing;
-            }
-            cr.fill().ok();
-        }
-    }
-
-    cr.restore().ok();
-    // Render cursor
-    if let Some((cursor_pos, cursor_shape)) = &rw.cursor {
-        if let Some(rl) = rw.lines.get(cursor_pos.view_line) {
-            layout.set_text(&rl.raw_text);
-            let cursor_attrs = build_pango_attrs(&rl.spans);
-            layout.set_attributes(Some(&cursor_attrs));
-
-            // When Ctrl+D selections are active, draw bar at right edge (col+1)
-            let render_col = if !rw.extra_selections.is_empty() && *cursor_shape == CursorShape::Bar
-            {
-                cursor_pos.col + 1
-            } else {
-                cursor_pos.col
-            };
-            let byte_offset: usize = rl
-                .raw_text
-                .char_indices()
-                .nth(render_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-
-            let pos = layout.index_to_pos(byte_offset as i32);
-            let cursor_x = text_x_offset + pos.x() as f64 / pango::SCALE as f64;
-            let char_w = pos.width() as f64 / pango::SCALE as f64;
-            let cursor_y = rect.y + cursor_pos.view_line as f64 * line_height;
-
-            let (cr_r, cr_g, cr_b) = theme.cursor.to_cairo();
-            let char_w = if char_w > 0.0 {
-                char_w
-            } else {
-                font_metrics.approximate_char_width() as f64 / pango::SCALE as f64
-            };
-            match cursor_shape {
-                CursorShape::Block => {
-                    cr.set_source_rgba(cr_r, cr_g, cr_b, theme.cursor_normal_alpha);
-                    cr.rectangle(cursor_x, cursor_y, char_w, line_height);
-                    cr.fill().ok();
-                }
-                CursorShape::Bar => {
-                    cr.set_source_rgb(cr_r, cr_g, cr_b);
-                    cr.rectangle(cursor_x, cursor_y, 2.0, line_height);
-                    cr.fill().ok();
-                }
-                CursorShape::Underline => {
-                    cr.set_source_rgb(cr_r, cr_g, cr_b);
-                    let bar_h = (line_height * 0.12).max(2.0);
-                    cr.rectangle(cursor_x, cursor_y + line_height - bar_h, char_w, bar_h);
-                    cr.fill().ok();
-                }
-            }
-        }
-    }
-
-    // AI ghost text — draw after cursor so it's clearly a suggestion.
-    if let Some((cursor_pos, _)) = &rw.cursor {
-        if let Some(rl) = rw.lines.get(cursor_pos.view_line) {
-            if let Some(ghost) = &rl.ghost_suffix {
-                layout.set_text(&rl.raw_text);
-                let ghost_line_attrs = build_pango_attrs(&rl.spans);
-                layout.set_attributes(Some(&ghost_line_attrs));
-                let byte_offset: usize = rl
-                    .raw_text
-                    .char_indices()
-                    .nth(cursor_pos.col)
-                    .map(|(i, _)| i)
-                    .unwrap_or(rl.raw_text.len());
-                let pos = layout.index_to_pos(byte_offset as i32);
-                let ghost_x = text_x_offset + pos.x() as f64 / pango::SCALE as f64;
-                let ghost_y = rect.y + cursor_pos.view_line as f64 * line_height;
-                let (gr, gg, gb) = theme.ghost_text_fg.to_cairo();
-                cr.set_source_rgb(gr, gg, gb);
-                cr.move_to(ghost_x, ghost_y);
-                layout.set_text(ghost);
-                layout.set_attributes(None);
-                pangocairo::show_layout(cr, layout);
-            }
-        }
-    }
-
-    // Secondary cursors — same shape as primary cursor (bar in Insert/VSCode, block in Normal).
-    let extra_cursor_shape = rw
-        .cursor
-        .as_ref()
-        .map(|(_, s)| *s)
-        .unwrap_or(CursorShape::Bar);
-    let has_extra_sels = !rw.extra_selections.is_empty();
-    let fallback_char_w = font_metrics.approximate_char_width() as f64 / pango::SCALE as f64;
-    let (cr_r, cr_g, cr_b) = theme.cursor.to_cairo();
-    for extra_pos in &rw.extra_cursors {
-        if let Some(rl) = rw.lines.get(extra_pos.view_line) {
-            layout.set_text(&rl.raw_text);
-            let extra_attrs = build_pango_attrs(&rl.spans);
-            layout.set_attributes(Some(&extra_attrs));
-            // When Ctrl+D selections are active, draw bar at right edge (col+1)
-            let render_col = if has_extra_sels && extra_cursor_shape == CursorShape::Bar {
-                extra_pos.col + 1
-            } else {
-                extra_pos.col
-            };
-            let byte_offset: usize = rl
-                .raw_text
-                .char_indices()
-                .nth(render_col)
-                .map(|(i, _)| i)
-                .unwrap_or(rl.raw_text.len());
-            let pos = layout.index_to_pos(byte_offset as i32);
-            let ex = text_x_offset + pos.x() as f64 / pango::SCALE as f64;
-            let ew = {
-                let w = pos.width() as f64 / pango::SCALE as f64;
-                if w > 0.0 {
-                    w
-                } else {
-                    fallback_char_w
-                }
-            };
-            let ey = rect.y + extra_pos.view_line as f64 * line_height;
-            match extra_cursor_shape {
-                CursorShape::Bar => {
-                    cr.set_source_rgb(cr_r, cr_g, cr_b);
-                    cr.rectangle(ex, ey, 2.0, line_height);
-                    cr.fill().ok();
-                }
-                CursorShape::Block => {
-                    cr.set_source_rgba(cr_r, cr_g, cr_b, theme.cursor_normal_alpha);
-                    cr.rectangle(ex, ey, ew, line_height);
-                    cr.fill().ok();
-                }
-                CursorShape::Underline => {
-                    cr.set_source_rgb(cr_r, cr_g, cr_b);
-                    let bar_h = (line_height * 0.12).max(2.0);
-                    cr.rectangle(ex, ey + line_height - bar_h, ew, bar_h);
-                    cr.fill().ok();
-                }
-            }
-        }
-    }
-}
-
-/// Convert a slice of [`StyledSpan`]s into a Pango [`AttrList`].
-pub(super) fn build_pango_attrs(spans: &[StyledSpan]) -> AttrList {
-    let attrs = AttrList::new();
-    for span in spans {
-        let (fr, fg_g, fb) = span.style.fg.to_pango_u16();
-        let mut fg_attr = AttrColor::new_foreground(fr, fg_g, fb);
-        fg_attr.set_start_index(span.start_byte as u32);
-        fg_attr.set_end_index(span.end_byte as u32);
-        attrs.insert(fg_attr);
-
-        if let Some(bg) = span.style.bg {
-            let (br, bg_g, bb) = bg.to_pango_u16();
-            let mut bg_attr = AttrColor::new_background(br, bg_g, bb);
-            bg_attr.set_start_index(span.start_byte as u32);
-            bg_attr.set_end_index(span.end_byte as u32);
-            attrs.insert(bg_attr);
-        }
-        if span.style.bold {
-            let mut w = pango::AttrInt::new_weight(pango::Weight::Bold);
-            w.set_start_index(span.start_byte as u32);
-            w.set_end_index(span.end_byte as u32);
-            attrs.insert(w);
-        }
-        if span.style.italic {
-            let mut s = pango::AttrInt::new_style(pango::Style::Italic);
-            s.set_start_index(span.start_byte as u32);
-            s.set_end_index(span.end_byte as u32);
-            attrs.insert(s);
-        }
-        if (span.style.font_scale - 1.0).abs() > f64::EPSILON {
-            let mut sc = pango::AttrFloat::new_scale(span.style.font_scale);
-            sc.set_start_index(span.start_byte as u32);
-            sc.set_end_index(span.end_byte as u32);
-            attrs.insert(sc);
-        }
-    }
-    attrs
-}
-
-#[allow(clippy::too_many_arguments)]
-pub(super) fn draw_visual_selection(
-    cr: &Context,
-    layout: &pango::Layout,
-    sel: &SelectionRange,
-    lines: &[render::RenderedLine],
-    rect: &WindowRect,
-    line_height: f64,
-    _scroll_top: usize,
-    text_x_offset: f64,
-    color: render::Color,
-    alpha: f64,
-) {
-    let (sr, sg, sb) = color.to_cairo();
-    cr.set_source_rgba(sr, sg, sb, alpha);
-
-    // Build a mapping from buffer line → view row index.  For each buffer
-    // line we take the LAST non-skippable rendered row (skipping wrap
-    // continuations, diff padding, and ghost continuations that share the
-    // same line_idx but occupy preceding view rows).
-    let mut line_to_view: std::collections::HashMap<usize, usize> =
-        std::collections::HashMap::new();
-    for (view_idx, rl) in lines.iter().enumerate() {
-        if rl.is_wrap_continuation || rl.is_ghost_continuation {
-            continue;
-        }
-        if rl.diff_status == Some(crate::core::engine::DiffLine::Padding) {
-            continue;
-        }
-        // Overwrite with latest view_idx → picks the content row, not a
-        // preceding padding/ghost row with the same line_idx.
-        line_to_view.insert(rl.line_idx, view_idx);
-    }
-    match sel.kind {
-        SelectionKind::Line => {
-            // Line-mode: highlight ALL visual rows (including wrap continuations)
-            // for each selected buffer line, so wrapped lines are fully covered.
-            for (view_idx, rl) in lines.iter().enumerate() {
-                if rl.line_idx >= sel.start_line
-                    && rl.line_idx <= sel.end_line
-                    && rl.diff_status != Some(crate::core::engine::DiffLine::Padding)
-                    && !rl.is_ghost_continuation
-                {
-                    let y = rect.y + view_idx as f64 * line_height;
-                    let highlight_width = rect.width - (text_x_offset - rect.x);
-                    cr.rectangle(text_x_offset, y, highlight_width, line_height);
-                }
-            }
-            cr.fill().ok();
-        }
-        SelectionKind::Char => {
-            for line_idx in sel.start_line..=sel.end_line {
-                let Some(&view_idx) = line_to_view.get(&line_idx) else {
-                    continue;
-                };
-                let rl = &lines[view_idx];
-                let y = rect.y + view_idx as f64 * line_height;
-                let line_text = &rl.raw_text;
-
-                layout.set_text(line_text);
-                layout.set_attributes(None);
-
-                if sel.start_line == sel.end_line {
-                    // Single-line selection
-                    let start_byte = line_text
-                        .char_indices()
-                        .nth(sel.start_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let start_pos = layout.index_to_pos(start_byte as i32);
-                    let start_x = text_x_offset + start_pos.x() as f64 / pango::SCALE as f64;
-
-                    let end_col = (sel.end_col + 1).min(line_text.chars().count());
-                    let end_byte = line_text
-                        .char_indices()
-                        .nth(end_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let end_pos = layout.index_to_pos(end_byte as i32);
-                    let end_x = text_x_offset + end_pos.x() as f64 / pango::SCALE as f64;
-
-                    cr.rectangle(start_x, y, end_x - start_x, line_height);
-                    cr.fill().ok();
-                } else if line_idx == sel.start_line {
-                    let start_byte = line_text
-                        .char_indices()
-                        .nth(sel.start_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let start_pos = layout.index_to_pos(start_byte as i32);
-                    let start_x = text_x_offset + start_pos.x() as f64 / pango::SCALE as f64;
-                    let (line_width, _) = layout.pixel_size();
-                    cr.rectangle(
-                        start_x,
-                        y,
-                        text_x_offset + line_width as f64 - start_x,
-                        line_height,
-                    );
-                    cr.fill().ok();
-                } else if line_idx == sel.end_line {
-                    let end_col = (sel.end_col + 1).min(line_text.chars().count());
-                    let end_byte = line_text
-                        .char_indices()
-                        .nth(end_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let end_pos = layout.index_to_pos(end_byte as i32);
-                    let end_x = text_x_offset + end_pos.x() as f64 / pango::SCALE as f64;
-                    cr.rectangle(text_x_offset, y, end_x - text_x_offset, line_height);
-                    cr.fill().ok();
-                } else {
-                    let (line_width, _) = layout.pixel_size();
-                    cr.rectangle(text_x_offset, y, line_width as f64, line_height);
-                    cr.fill().ok();
-                }
-            }
-        }
-        SelectionKind::Block => {
-            for line_idx in sel.start_line..=sel.end_line {
-                let Some(&view_idx) = line_to_view.get(&line_idx) else {
-                    continue;
-                };
-                let rl = &lines[view_idx];
-                let y = rect.y + view_idx as f64 * line_height;
-                let line_text = &rl.raw_text;
-                let line_len = line_text.chars().count();
-
-                layout.set_text(line_text);
-                layout.set_attributes(None);
-
-                if sel.start_col < line_len {
-                    let start_byte = line_text
-                        .char_indices()
-                        .nth(sel.start_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let start_pos = layout.index_to_pos(start_byte as i32);
-                    let start_x = text_x_offset + start_pos.x() as f64 / pango::SCALE as f64;
-
-                    let block_end_col = (sel.end_col + 1).min(line_len);
-                    let end_byte = line_text
-                        .char_indices()
-                        .nth(block_end_col)
-                        .map(|(i, _)| i)
-                        .unwrap_or(line_text.len());
-                    let end_pos = layout.index_to_pos(end_byte as i32);
-                    let end_x = text_x_offset + end_pos.x() as f64 / pango::SCALE as f64;
-
-                    cr.rectangle(start_x, y, end_x - start_x, line_height);
-                }
-            }
-            cr.fill().ok();
-        }
-    }
 }
 
 pub(super) fn draw_window_separators(
